requirements.freeze.txt
amqp==2.2.2
apipkg==1.4
argh==0.26.2
attrs==17.4.0
autopep8==1.3.2
backports-abc==0.5
billiard==3.5.0.3
bleach==2.1.2
boto3==1.4.7
botocore==1.7.48
celery==4.2.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
decorator==4.2.1
Django==1.11.5
django-cors-headers==2.1.0
django-enumfield==1.2.1
django-extensions==1.9.1
django-extras==0.3
django-filter==1.0.4
django-redis==4.8.0
django-rest-swagger==2.1.2
django-storages==1.6.5
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
docutils==0.14
drf-nested-routers==0.90.0
elasticsearch==5.4.0
entrypoints==0.2.3
execnet==1.5.0
factory-boy==2.9.2
Faker==0.8.11
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
kombu==4.2.1
livereload==2.5.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
mkdocs==0.16.3
mockredispy==2.9.3
msgpack-python==0.4.8
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.3.1
olefile==0.45.1
openapi-codec==1.3.2
pandocfilters==1.4.2
parso==0.1.1
pathtools==0.1.2
pep8==1.7.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.2.1
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.5.3
pytest==3.3.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-forked==0.2
pytest-xdist==1.20.1
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
Send2Trash==1.4.2
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.1
tornado==4.5.3
traitlets==4.3.2
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.3
